+++ config/easy/Mirrors/code/CGCode.js
@@ -1,0 +1,7 @@
+const n = parseInt(readline());
+const r = readline().split(' ').map(j => parseFloat(j)).slice(0, n);
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('reflected light');