+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - configuration for "Video Assistant Referee".
 - configuration for "Ulam Spiral".
 - configuration for "The leaking bathtub".
+- configuration for "Mirrors".
 
 ### Fixed
 - alphanumName configuration for "Folding a Note".